[utils] Github, organization, repository service class 분리
