+++ opgc/conf/settings/base.py
@@ -19,6 +19,3 @@ INSTALLED_APPS = [
 
 # SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = 'OPGC_SECRET_KEY'
-
-# Gihub api auth token
-OPGC_TOKEN = 'OPGG_GITHUB_TOKEN'
