+++ opgc/conf/settings/prod.py
@@ -13,6 +13,11 @@ DEBUG = True
 IS_PROD = True
 
 ALLOWED_HOSTS = ['OPGC_ALLOWED_HOSTS']
+
+# Gihub api auth token
+OPGC_TOKEN = 'OPGG_GITHUB_TOKEN'
+GITHUB_API_HEADER = {'Authorization': f'token {OPGC_TOKEN}'}
+
 # 임시
 CORS_ORIGIN_ALLOW_ALL = True
 # CORS_ALLOW_CREDENTIALS = True
